qgspointcloudrequest: Change setFilterRect to setFilterGeometry
The `setFilterRect` allows to filter a pointcloud request by a
rectangular extent but it does not allow to set any geometry. The
change to `setFilterGeometry` allows to specify any geometry.

The change in the decoder are minimal. They still check is inside the
filter geometry.

If the requested filter is a QgsRectangle, it can be transformed into
a Geometry by using `QgsGeometry::fromRect`.